backend/app/app/requirements.txt
SQLAlchemy==1.4.3
psycopg2-binary==2.8.6
alembic==1.5.8
tenacity==7.0.0
pytest==6.2.2
pytest-asyncio==0.14.0
pytest-mock==3.5.1
requests==2.25.1
mock==4.0.3
mypy==0.812
isort==5.8.0
flake8==3.9.0
